refactor(dto): Change CustomerSegment enum in Customer entity
