+++ services/inventory/src/main/java/com/petek/inventoryService/dto/CustomerResponse.java
@@ -2,7 +2,7 @@ package com.petek.inventoryService.dto;
 
 import java.time.Instant;
 
-import com.petek.inventoryService.entity.CustomerSegment;
+import com.petek.inventoryService.entity.Customer.CustomerSegment;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
