+++ services/inventory/src/main/java/com/petek/inventoryService/dto/ProductFilterRequest.java
@@ -18,7 +18,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductFilterRequest {
-    
     @NotNull
     @Min(0)
     @Builder.Default
