--- services/inventory/src/main/java/com/petek/inventoryService/entity/CustomerSegment.java
@@ -1,5 +1,0 @@
-package com.petek.inventoryService.entity;
-
-public enum CustomerSegment {
-    INDIVIDUAL, SME, CORPORATE, ENTERPRISE, OTHER
-}
